requirements.txt
backcall==0.2.0
black==21.12b0
click==8.0.3
colorama==0.4.4
cv==1.0.0
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.3
ipykernel==6.6.1
ipython==7.31.0
jedi==0.18.1
jupyter-client==7.1.0
jupyter-core==4.9.1
matplotlib-inline==0.1.3
MouseInfo==0.1.3
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numpy==1.22.0
opencv-python==4.5.5.62
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
prompt-toolkit==3.0.24
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
Pygments==2.11.2
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.4
pywin32==303
pyzmq==22.3.0
six==1.16.0
tomli==1.2.3
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
wcwidth==0.2.5
